requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.41.0
numpy>=1.24.0
pyyaml>=6.0.1
tqdm>=4.66.0
Pillow>=9.5.0
lightning>=2.2.0
hydra-core>=1.3.2
timm>=1.0.7
omegaconf>=2.3.0

